[Perf] Split Qwen VL and Mamba memcpy warmup out of GDN warmup
