+++ tests/model_executor/test_qwen_triton_warmup.py
@@ -6,16 +6,12 @@ from types import SimpleNamespace
 import pytest
 import torch
 
-import vllm.model_executor.warmup.qwen_triton_warmup as warmup_module
 from vllm.model_executor.warmup.qwen_triton_warmup import (
     _FLA_POST_CONV_WARMUP_LENGTHS,
     _qwen_gdn_warmup_config,
-    _warm_batch_memcpy_kernel,
     _warm_causal_conv1d_fwd_kernel,
     _warm_fused_post_conv_kernel,
     _warm_gated_rms_norm_kernel,
-    _warm_mrope,
-    _warm_vision,
     qwen_triton_warmup,
 )
 
@@ -60,6 +56,7 @@ def _model_config(model_type: str = "qwen3_5") -> SimpleNamespace:
     return SimpleNamespace(
         hf_text_config=SimpleNamespace(model_type=model_type),
         hf_config=SimpleNamespace(model_type=model_type),
+        dtype=torch.bfloat16,
     )
 
 
@@ -70,7 +67,7 @@ def _stub_qwen_warmup_helpers(monkeypatch, calls: list[str]) -> None:
     )
     monkeypatch.setattr(
         "vllm.model_executor.warmup.qwen_triton_warmup._warm_gated_rms_norm_kernel",
-        lambda device, value, max_num_tokens: calls.append("rmsnorm"),
+        lambda device, value, max_num_tokens, x_dtype=None: calls.append("rmsnorm"),
     )
     monkeypatch.setattr(
         "vllm.model_executor.warmup.qwen_triton_warmup._warm_causal_conv1d_fwd_kernel",
@@ -81,21 +78,9 @@ def _stub_qwen_warmup_helpers(monkeypatch, calls: list[str]) -> None:
         lambda device, value: calls.append("post"),
     )
     monkeypatch.setattr(
-        "vllm.model_executor.warmup.qwen_triton_warmup._warm_batch_memcpy_kernel",
-        lambda device: calls.append("memcpy"),
-    )
-    monkeypatch.setattr(
         "vllm.model_executor.warmup.qwen_triton_warmup"
         "._warm_fused_sigmoid_gating_delta_rule_update_kernel",
         lambda device, value: calls.append("decode"),
-    )
-    monkeypatch.setattr(
-        "vllm.model_executor.warmup.qwen_triton_warmup._warm_vision",
-        lambda model: calls.append("vision"),
-    )
-    monkeypatch.setattr(
-        "vllm.model_executor.warmup.qwen_triton_warmup._warm_mrope",
-        lambda runner, model: calls.append("mrope"),
     )
     monkeypatch.setattr(
         "vllm.model_executor.warmup.qwen_triton_warmup._synchronize_device",
@@ -142,18 +127,9 @@ def test_qwen_triton_warmup_runs_prefill_kernels_for_pooling(monkeypatch) -> Non
         device=torch.device("cpu"),
         max_num_tokens=512,
         compilation_config=SimpleNamespace(static_forward_context={}),
-        get_model=lambda: torch.nn.Module(),
     )
     qwen_triton_warmup(runner, _model_config())
-    assert calls == [
-        "rmsnorm",
-        "conv",
-        "post",
-        "memcpy",
-        "vision",
-        "mrope",
-        "sync",
-    ]
+    assert calls == ["rmsnorm", "conv", "post", "sync"]
 
 
 def test_qwen_triton_warmup_runs_generate_kernels(monkeypatch) -> None:
@@ -164,19 +140,9 @@ def test_qwen_triton_warmup_runs_generate_kernels(monkeypatch) -> None:
         device=torch.device("cpu"),
         max_num_tokens=512,
         compilation_config=SimpleNamespace(static_forward_context={}),
-        get_model=lambda: torch.nn.Module(),
     )
     qwen_triton_warmup(runner, _model_config())
-    assert calls == [
-        "rmsnorm",
-        "conv",
-        "post",
-        "memcpy",
-        "decode",
-        "vision",
-        "mrope",
-        "sync",
-    ]
+    assert calls == ["rmsnorm", "conv", "post", "decode", "sync"]
 
 
 def test_qwen_triton_warmup_skips_non_qwen_model_type(monkeypatch) -> None:
@@ -187,42 +153,31 @@ def test_qwen_triton_warmup_skips_non_qwen_model_type(monkeypatch) -> None:
         "vllm.model_executor.warmup.qwen_triton_warmup._qwen_gdn_warmup_config",
         fail,
     )
-    runner = SimpleNamespace(is_pooling_model=False, get_model=fail)
+    runner = SimpleNamespace(is_pooling_model=False)
     qwen_triton_warmup(runner, _model_config("custom"))
 
 
-def test_qwen_triton_warmup_runs_vl_helpers_when_gdn_config_missing(
-    monkeypatch,
-) -> None:
-    calls = []
-    model = torch.nn.Module()
+def test_qwen_triton_warmup_skips_when_gdn_config_missing(monkeypatch) -> None:
+    def fail(*_args, **_kwargs):
+        raise AssertionError("GDN kernels must not run without a GDN config")
+
+    monkeypatch.setattr(
+        "vllm.model_executor.warmup.qwen_triton_warmup._qwen_gdn_warmup_config",
+        lambda *args, **kwargs: None,
+    )
+    monkeypatch.setattr(
+        "vllm.model_executor.warmup.qwen_triton_warmup._warm_gated_rms_norm_kernel",
+        fail,
+    )
     runner = SimpleNamespace(
-        device=torch.device("cuda"),
         is_pooling_model=False,
+        device=torch.device("cpu"),
         compilation_config=SimpleNamespace(static_forward_context={}),
-        get_model=lambda: model,
     )
-
-    monkeypatch.setattr(
-        warmup_module, "_qwen_gdn_warmup_config", lambda *args, **kwargs: None
-    )
-    monkeypatch.setattr(
-        warmup_module, "_warm_vision", lambda value: calls.append(value)
-    )
-    monkeypatch.setattr(
-        warmup_module,
-        "_warm_mrope",
-        lambda runner_value, model_value: calls.append((runner_value, model_value)),
-    )
-    monkeypatch.setattr(
-        warmup_module, "_synchronize_device", lambda device: calls.append(device)
-    )
-
     qwen_triton_warmup(runner, _model_config())
-    assert calls == [model, (runner, model), runner.device]
 
 
-def test_warm_gated_rms_norm_uses_production_m_shape(monkeypatch) -> None:
+def test_warm_gated_rms_norm_uses_production_shape(monkeypatch) -> None:
     captured: list[dict[str, object]] = []
 
     def fake_warmup(**kwargs):
@@ -233,10 +188,7 @@ def test_warm_gated_rms_norm_uses_production_m_shape(monkeypatch) -> None:
         ".warmup_layer_norm_fwd",
         fake_warmup,
     )
-    from vllm.model_executor.warmup.qwen_triton_warmup import (
-        _QwenGDNWarmupConfig,
-        _warm_gated_rms_norm_kernel,
-    )
+    from vllm.model_executor.warmup.qwen_triton_warmup import _QwenGDNWarmupConfig
 
     config = _QwenGDNWarmupConfig(
         h=2,
@@ -328,107 +280,10 @@ def _cuda_gdn_config() -> object:
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA is required")
 def test_qwen_gdn_prefill_warmup_kernels_compile_on_gpu() -> None:
-    from vllm.model_executor.warmup.qwen_triton_warmup import (
-        _warm_gated_rms_norm_kernel,
-    )
-
     config = _cuda_gdn_config()
     device = torch.device("cuda")
     _warm_gated_rms_norm_kernel(device, config, max_num_tokens=16)
     _warm_causal_conv1d_fwd_kernel(device, config)
     _warm_fused_post_conv_kernel(device, config)
-    _warm_batch_memcpy_kernel(device)
     assert _FLA_POST_CONV_WARMUP_LENGTHS == (1, 2, 16)
     torch.accelerator.synchronize(device)
-
-
-def test_vision_warmup_calls_only_position_and_rotary_paths() -> None:
-    from vllm.model_executor.models.qwen3_vl import Qwen3_VisionTransformer
-
-    calls: list[tuple[str, object]] = []
-
-    class FakeAttention:
-        num_attention_heads_per_partition = 4
-        hidden_size_per_attention_head = 8
-
-        def apply_rotary_emb(self, qk, cos, sin):
-            calls.append(("rotary", qk.shape))
-
-    class FakeVisual(Qwen3_VisionTransformer):
-        spatial_merge_size = 2
-
-        def __init__(self) -> None:
-            torch.nn.Module.__init__(self)
-            self.blocks = [SimpleNamespace(attn=FakeAttention())]
-
-        def fast_pos_embed_interpolate(self, grid_thw):
-            calls.append(("position", grid_thw[0]))
-
-        def rot_pos_emb(self, grid_thw):
-            _, h, w = grid_thw[0]
-            shape = (h * w, 4)
-            return torch.empty(shape), torch.empty(shape)
-
-        def forward(self, *args, **kwargs):
-            raise AssertionError("vision warmup must not run the full tower")
-
-    model = torch.nn.Module()
-    model.visual = FakeVisual()
-    _warm_vision(model)
-
-    grids = [(1, 16, 16), (1, 16, 2), (1, 2, 16), (1, 2, 2)]
-    assert [value for name, value in calls if name == "position"] == [
-        list(grid) for grid in grids
-    ]
-    assert [value for name, value in calls if name == "rotary"] == [
-        torch.Size((2, h * w, 4, 8)) for _, h, w in grids
-    ]
-
-
-def test_mrope_warmup_reads_model_config_on_v2_runner() -> None:
-    from vllm.model_executor.layers.rotary_embedding.mrope import MRotaryEmbedding
-
-    launched: list[tuple[torch.Size, torch.Size]] = []
-
-    class FakeRope(MRotaryEmbedding):
-        def __init__(self) -> None:
-            torch.nn.Module.__init__(self)
-            self.head_size = 8
-            self.rotary_dim = 8
-            self.mrope_section = [2, 3, 3]
-            self.mrope_interleaved = False
-            self.is_neox_style = True
-
-        def forward(self, positions, query, key):
-            launched.append((positions.shape, query.shape))
-            return query, key
-
-    model = torch.nn.Module()
-    model.rotary_emb = FakeRope()
-    runner = SimpleNamespace(
-        model_config=SimpleNamespace(
-            uses_mrope=True,
-            get_num_attention_heads=lambda parallel_config: 4,
-            get_num_kv_heads=lambda parallel_config: 2,
-        ),
-        parallel_config=object(),
-        dtype=torch.bfloat16,
-        device=torch.device("cpu"),
-    )
-    _warm_mrope(runner, model)
-    assert [shape for shape, _ in launched] == [
-        torch.Size((3, 1)),
-        torch.Size((3, 2)),
-        torch.Size((3, 16)),
-    ]
-
-
-def test_mrope_warmup_skips_when_model_config_disables_mrope() -> None:
-    def fail(*_args, **_kwargs):
-        raise AssertionError("M-RoPE warmup must not run when uses_mrope is false")
-
-    runner = SimpleNamespace(
-        model_config=SimpleNamespace(uses_mrope=False),
-        get_model=fail,
-    )
-    _warm_mrope(runner, torch.nn.Module())
