+++ vllm/model_executor/warmup/qwen_triton_warmup.py
@@ -8,7 +8,6 @@ from typing import TYPE_CHECKING
 import torch
 
 from vllm.logger import init_logger
-from vllm.utils.math_utils import round_up
 
 if TYPE_CHECKING:
     from vllm.config import ModelConfig
@@ -28,7 +27,6 @@ _QWEN_MODEL_TYPES = frozenset(
 
 # Covers L=1 constexpr, non-divisible runtime L, and divisible runtime L.
 _FLA_POST_CONV_WARMUP_LENGTHS = (1, 2, 16)
-_MROPE_TOKEN_COUNTS = (1, 2, 16)
 
 
 @dataclass(frozen=True)
@@ -288,129 +286,9 @@ def _warm_fused_sigmoid_gating_delta_rule_update_kernel(
     )
 
 
-def _warm_batch_memcpy_kernel(device: torch.device) -> None:
-    """Warm the Mamba/GDN prefix-cache state copy specialization."""
-    if device.type != "cuda":
-        return
-
-    from vllm.v1.worker.mamba_utils import batch_memcpy
-
-    src = torch.empty(1024, dtype=torch.uint8, device=device)
-    dst = torch.empty_like(src)
-    batch_memcpy(
-        torch.tensor([src.data_ptr()], dtype=torch.uint64, device=device),
-        torch.tensor([dst.data_ptr()], dtype=torch.uint64, device=device),
-        # Keep this int32: Triton's compile key includes pointer element dtypes,
-        # and the production prefix-cache path passes an int32 sizes tensor.
-        torch.tensor([src.numel()], dtype=torch.int32, device=device),
-    )
-    logger.info("Warmed Mamba/GDN batch_memcpy_kernel.")
-
-
 def _synchronize_device(device: torch.device) -> None:
     if device.type == "cuda":
         torch.accelerator.synchronize(device)
-
-
-def _warm_vision(model: torch.nn.Module) -> None:
-    from vllm.model_executor.models.qwen3_vl import Qwen3_VisionTransformer
-
-    for visual in model.modules():
-        if not isinstance(visual, Qwen3_VisionTransformer):
-            continue
-        attention = visual.blocks[0].attn
-        num_heads = int(attention.num_attention_heads_per_partition)
-        head_size = int(attention.hidden_size_per_attention_head)
-        merge_size = int(visual.spatial_merge_size)
-        divisible = round_up(16, merge_size)
-        while divisible % 16:
-            divisible += merge_size
-        spatial_sizes = [divisible]
-        if merge_size % 16:
-            spatial_sizes.append(merge_size)
-            if merge_size == 1:
-                spatial_sizes.append(2)
-        grids = [(1, h, w) for h in spatial_sizes for w in spatial_sizes]
-
-        for grid in grids:
-            grid_thw = [list(grid)]
-            visual.fast_pos_embed_interpolate(grid_thw)
-            cos, sin = visual.rot_pos_emb(grid_thw)
-            qk = torch.empty(
-                (2, grid[0] * grid[1] * grid[2], num_heads, head_size),
-                dtype=cos.dtype,
-                device=cos.device,
-            )
-            attention.apply_rotary_emb(qk, cos, sin)
-
-        logger.info(
-            "Warmed position embedding and vision rotary kernels on grids=%s.",
-            grids,
-        )
-
-
-def _runner_uses_mrope(runner: object) -> bool:
-    uses_mrope = getattr(runner, "uses_mrope", None)
-    if uses_mrope is not None:
-        return bool(uses_mrope)
-    model_config = getattr(runner, "model_config", None)
-    return bool(getattr(model_config, "uses_mrope", False))
-
-
-def _runner_num_query_heads(runner: object) -> int:
-    num_query_heads = getattr(runner, "num_query_heads", None)
-    if num_query_heads is not None:
-        return int(num_query_heads)
-    model_config = getattr(runner, "model_config", None)
-    parallel_config = getattr(runner, "parallel_config", None)
-    return int(model_config.get_num_attention_heads(parallel_config))
-
-
-def _warm_mrope(runner: "GPUModelRunner", model: torch.nn.Module) -> None:
-    if not _runner_uses_mrope(runner):
-        return
-
-    from vllm.model_executor.layers.rotary_embedding.mrope import MRotaryEmbedding
-
-    num_query_heads = _runner_num_query_heads(runner)
-    num_kv_heads = int(runner.model_config.get_num_kv_heads(runner.parallel_config))
-    seen: set[tuple[object, ...]] = set()
-
-    for rope in model.modules():
-        if not isinstance(rope, MRotaryEmbedding):
-            continue
-        key = (
-            rope.head_size,
-            rope.rotary_dim,
-            tuple(rope.mrope_section or ()),
-            rope.mrope_interleaved,
-            rope.is_neox_style,
-            runner.dtype,
-        )
-        if key in seen:
-            continue
-        seen.add(key)
-
-        for num_tokens in _MROPE_TOKEN_COUNTS:
-            positions = torch.arange(
-                num_tokens, dtype=torch.long, device=runner.device
-            ).expand(3, -1)
-            query = torch.empty(
-                (num_tokens, num_query_heads * rope.head_size),
-                dtype=runner.dtype,
-                device=runner.device,
-            )
-            key_tensor = torch.empty(
-                (num_tokens, num_kv_heads * rope.head_size),
-                dtype=runner.dtype,
-                device=runner.device,
-            )
-            rope(positions, query, key_tensor)
-
-    if seen:
-        logger.info(
-            "Warmed M-RoPE Triton kernels for %d compile-key shape(s).", len(seen)
-        )
 
 
 @torch.inference_mode()
@@ -418,7 +296,7 @@ def qwen_triton_warmup(
     runner: "GPUModelRunner",
     model_config: "ModelConfig",
 ) -> None:
-    """Warm Qwen Triton kernels reported by the JIT monitor."""
+    """Warm Qwen GDN Triton kernels reported by the JIT monitor."""
     hf_text_config = getattr(model_config, "hf_text_config", None)
     hf_config = getattr(model_config, "hf_config", None)
     model_type = None
@@ -431,23 +309,20 @@ def qwen_triton_warmup(
         return
 
     device = getattr(runner, "device", torch.device("cuda"))
-    logger.info("Warming up Qwen Triton kernels for model_type=%s.", model_type)
+    logger.info("Warming up Qwen GDN Triton kernels for model_type=%s.", model_type)
 
     compilation_config = getattr(runner, "compilation_config", None)
     static_forward_context = getattr(compilation_config, "static_forward_context", None)
     gdn_config = _qwen_gdn_warmup_config(static_forward_context)
-    if gdn_config is not None:
-        max_num_tokens = max(1, int(getattr(runner, "max_num_tokens", 1)))
-        x_dtype = getattr(model_config, "dtype", gdn_config.conv_dtype)
-        _warm_gated_rms_norm_kernel(device, gdn_config, max_num_tokens, x_dtype)
-        _warm_causal_conv1d_fwd_kernel(device, gdn_config)
-        _warm_fused_post_conv_kernel(device, gdn_config)
-        _warm_batch_memcpy_kernel(device)
-        # Pooling only runs full prefills; the decode update kernel is unused.
-        if not runner.is_pooling_model:
-            _warm_fused_sigmoid_gating_delta_rule_update_kernel(device, gdn_config)
+    if gdn_config is None:
+        return
 
-    model = runner.get_model()
-    _warm_vision(model)
-    _warm_mrope(runner, model)
+    max_num_tokens = max(1, int(getattr(runner, "max_num_tokens", 1)))
+    x_dtype = getattr(model_config, "dtype", None) or gdn_config.conv_dtype
+    _warm_gated_rms_norm_kernel(device, gdn_config, max_num_tokens, x_dtype)
+    _warm_causal_conv1d_fwd_kernel(device, gdn_config)
+    _warm_fused_post_conv_kernel(device, gdn_config)
+    # Pooling only runs full prefills; the decode update kernel is unused.
+    if not runner.is_pooling_model:
+        _warm_fused_sigmoid_gating_delta_rule_update_kernel(device, gdn_config)
     _synchronize_device(device)
